refactor: 경매 상세조회 컴포넌트 분리 시키기
