+++ src/components/Auction/AuctionImage.tsx
@@ -1,0 +1,7 @@
+interface Props {
+  imageUrl: string;
+}
+
+export default function AuctionImage({ imageUrl }: Props) {
+  return <img src={imageUrl} className="w-full object-contain" alt="auction" />;
+}
